backend/requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
pydantic>=2.5.2
pydantic-settings>=2.1.0
pydantic[email]>=2.5.2
redis>=5.0.1
celery>=5.3.6
loguru>=0.7.2
python-dotenv>=1.0.0
emails>=0.6
jinja2>=3.1.2
pillow>=10.1.0
fpdf>=1.7.2
httpx>=0.25.2